warnings
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@11636 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/WorldBuilder.cxx
@@ -196,7 +196,7 @@ void* WorldBuilder::unpack(void* buf)
   
   if (debugLevel >= 3) {
     TimeKeeper end = TimeKeeper::getCurrent();
-    const float elapsed = end - start;
+    const float elapsed = (float)(end - start);
     DEBUG0("WorldBuilder::unpack() processed in %f seconds.\n", elapsed);
   }
